+++ netpyne/cell/pointCell.py
@@ -418,10 +418,12 @@ class PointCell (Cell):
 
     def __getattr__(self, name):
         def wrapper(*args, **kwargs):
+            from .. import sim
             try: 
                 name(*args,**kwargs)
             except:
-                print("Error: Function '%s' not yet implemented for Point Neurons" % name)
+                if sim.cfg.verbose:
+                    print("Error: Function '%s' not yet implemented for Point Neurons" % name)
         return wrapper
 
     # def modify (self):