+++ netpyne/sim/run.py
@@ -55,8 +55,8 @@ def preRun ():
     # handler for printing out time during simulation run
     if sim.rank == 0 and sim.cfg.printRunTime:
         def printRunTime():
-            for i in range(int(sim.cfg.printRunTime*1000.0), int(sim.cfg.duration), int(sim.cfg.printRunTime*1000.0)):
-                sim.cvode.event(i, 'print ' + str(i/1000.0) + ',"s"')
+            print('%.1gs' % (h.t/1000.0))   
+            sim.cvode.event(h.t + int(sim.cfg.printRunTime*1000.0), sim.printRunTime)
 
         sim.printRunTime = printRunTime
         sim.fih.append(h.FInitializeHandler(1, sim.printRunTime))
@@ -90,11 +90,12 @@ def preRun ():
     # handler for recording LFP
     if sim.cfg.recordLFP:
         def recordLFPHandler():
-            for i in np.arange(sim.cfg.recordStep, sim.cfg.duration+sim.cfg.recordStep, sim.cfg.recordStep):
-                sim.cvode.event(i, sim.calculateLFP)
+            sim.cvode.event(h.t + int(sim.cfg.recordStep), sim.calculateLFP)
+            sim.cvode.event(h.t + int(sim.cfg.recordStep), recordLFPHandler)
 
         sim.recordLFPHandler = recordLFPHandler
         sim.fih.append(h.FInitializeHandler(0, sim.recordLFPHandler))  # initialize imemb
+
 
 #------------------------------------------------------------------------------
 # Run Simulation
@@ -175,7 +176,7 @@ def calculateLFP():
 #------------------------------------------------------------------------------
 # Calculate and print load balance
 #------------------------------------------------------------------------------
-def loadBalance ():
+def loadBalance (printNodeTimes = False):
     from .. import sim
 
     computation_time = sim.pc.step_time()
@@ -184,7 +185,9 @@ def loadBalance ():
     avg_comp_time = sim.pc.allreduce(computation_time, 1)/sim.nhosts
     load_balance = avg_comp_time/max_comp_time
 
-    print('node:',sim.rank,' comp_time:',computation_time)
+    if printNodeTimes:
+        print('node:',sim.rank,' comp_time:',computation_time)
+    
     if sim.rank==0:
         print('max_comp_time:', max_comp_time)
         print('min_comp_time:', min_comp_time)